+++ build.gradle
@@ -20,7 +20,7 @@ buildscript {
 		jcenter()
     }
     dependencies {
-        classpath 'com.huawei.ohos:hap:2.4.2.7'
+        classpath 'com.huawei.ohos:hap:2.4.4.2'
         classpath 'com.huawei.ohos:decctest:1.0.0.7'
         classpath "org.sonarsource.scanner.gradle:sonarqube-gradle-plugin:3.3"
     }
