+++ entry/build.gradle
@@ -20,7 +20,7 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar', '*.har'])
     testImplementation 'junit:junit:4.13'
     ohosTestImplementation 'com.huawei.ohos.testkit:runner:1.0.0.100'
-    implementation project(':support')
+    implementation project(':stripedprocessbutton')
 }
 decc {
     supportType = ['html','xml']
