+++ entry/src/main/config.json
@@ -15,7 +15,7 @@
   "deviceConfig": {},
   "module": {
     "package": "com.github.nikartm.stripedprocessbutton",
-    "name": ".MyApplication",
+    "name": "com.github.nikartm.stripedprocessbutton.sample.MyApplication",
     "deviceType": [
       "phone","tablet","tv","wearable"
     ],
@@ -37,7 +37,7 @@
           }
         ],
         "orientation": "unspecified",
-        "name": "com.github.nikartm.stripedprocessbutton.MainAbility",
+        "name": "com.github.nikartm.stripedprocessbutton.sample.MainAbility",
         "icon": "$media:icon",
         "description": "$string:mainability_description",
         "label": "$string:app_name",
