+++ entry/src/ohosTest/java/com/github/nikartm/stripedprocessbutton/sample/StripedProcessButtonTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.nikartm.stripedprocessbutton;
+package com.github.nikartm.stripedprocessbutton.sample;
 
 import static ohos.agp.components.AttrHelper.getDensity;
 import static org.junit.Assert.assertEquals;
@@ -23,10 +23,10 @@ import ohos.aafwk.ability.delegation.AbilityDelegatorRegistry;
 import ohos.agp.components.AttrSet;
 import ohos.agp.utils.Color;
 import ohos.app.Context;
-import com.github.nikartm.support.AnimatedStripedDrawable;
-import com.github.nikartm.support.AttributeController;
-import com.github.nikartm.support.StripedDrawable;
-import com.github.nikartm.support.StripedProcessButton;
+import com.github.nikartm.stripedprocessbutton.AnimatedStripedDrawable;
+import com.github.nikartm.stripedprocessbutton.AttributeController;
+import com.github.nikartm.stripedprocessbutton.StripedDrawable;
+import com.github.nikartm.stripedprocessbutton.StripedProcessButton;
 import org.junit.Before;
 import org.junit.Test;
