+++ stripedprocessbutton/.gitignore
