+++ stripedprocessbutton/build.gradle
