+++ stripedprocessbutton/proguard-rules.pro
