+++ stripedprocessbutton/src/main/java/com/github/nikartm/stripedprocessbutton/constant/Constants.java
@@ -1,4 +1,4 @@
-package com.github.nikartm.support.constant;
+package com.github.nikartm.stripedprocessbutton.constant;
 
 import ohos.agp.utils.Color;
