+++ stripedprocessbutton/src/main/resources/base/element/string.json
